// src/store/home/types.ts
export enum Types {
  SAVE_DATA_CATEGORY = 'SAVE_DATA_CATEGORY',
  SAVE_ITEM_COURSE = 'SAVE_ITEM_COURSE',
  SAVE_UPDATE_PROFILE = 'SAVE_UPDATE_PROFILE',
  SAVE_DATA_PROVINCES = 'SAVE_DATA_PROVINCES',
  SAVE_DATA_LEVELS = 'SAVE_DATA_LEVELS',
  SAVE_DATA_DISTRICTS = 'SAVE_DATA_DISTRICTS',
  SAVE_DATA_SCHOOLS = 'SAVE_DATA_SCHOOLS',
  RESET = 'RESET',
  SAVE_LIST_NOTIFY_ALl = 'SAVE_LIST_NOTIFY_ALl',
  SAVE_NUM_NOTI_NOT_READ = 'SAVE_NUM_NOTI_NOT_READ',
  SAVE_NEWS_NOTI_DATA = 'SAVE_NEWS_NOTI_DATA',
  LOADING_NOTI_DATA = 'LOADING_NOTI_DATA',
  IS_SHOW_NEWS = 'IS_SHOW_NEWS',
  IS_SHOW_POSTS = 'IS_SHOW_POSTS',
  SAVE_POSTS_NOTI_DATA = 'SAVE_POSTS_NOTI_DATA',
  SAVE_LAST_LESSON = 'SAVE_LAST_LESSON',
  SAVE_SLIDE_SHOW = 'SAVE_SLIDE_SHOW',
  IS_SHOW_AFFILIATE = 'IS_SHOW_AFFILIATE'
}